chore(AppModule): reorganize file structure
